sys/vfs/hammer: Remove unnecessary casts
These are all int minus int where both ints are flush group #.

+++ sys/vfs/hammer/hammer_flusher.c
@@ -165,7 +165,7 @@ hammer_flusher_async_one(hammer_mount_t hmp)
 void
 hammer_flusher_wait(hammer_mount_t hmp, int seq)
 {
-	while ((int)(seq - hmp->flusher.done) > 0)
+	while (seq - hmp->flusher.done > 0)
 		tsleep(&hmp->flusher.done, 0, "hmrfls", 0);
 }
 
@@ -178,7 +178,7 @@ int
 hammer_flusher_running(hammer_mount_t hmp)
 {
 	int seq = hmp->flusher.next - 1;
-	if ((int)(seq - hmp->flusher.done) > 0)
+	if (seq - hmp->flusher.done > 0)
 		return(1);
 	return (0);
 }
@@ -355,7 +355,7 @@ hammer_flusher_flush(hammer_mount_t hmp, int *nomorep)
 		 * filesystem error occurred which forced the filesystem into
 		 * read-only mode.
 		 */
-		KKASSERT((int)(flg->seq - seq) > 0 || hmp->ronly >= 2);
+		KKASSERT(flg->seq - seq > 0 || hmp->ronly >= 2);
 		flg = NULL;
 	}
 
@@ -447,7 +447,7 @@ hammer_flusher_flush(hammer_mount_t hmp, int *nomorep)
 	 * it can no longer be reused.
 	 */
 	while ((resv = TAILQ_FIRST(&hmp->delay_list)) != NULL) {
-		if ((int)(resv->flush_group - seq) > 0)
+		if (resv->flush_group - seq > 0)
 			break;
 		hammer_reserve_clrdelay(hmp, resv);
 	}
